Remove Address Resolver related features from MTD build
