+++ src/core/thread/mle_router_ftd.hpp
@@ -52,6 +52,7 @@
 namespace ot {
 namespace Mle {
 
+class AddressResolver;
 class MeshForwarder;
 class NetworkDataLeader;
